Revert "moved AbstractUIBasedConfigurationBuilder to de.uka.ipd.sdq.workflow.launchconfig.core"
This reverts commit fee9d8b6744832c3053915a40b8255a30d59cd53.
