+++ bundles/de.uka.ipd.sdq.workflow.launchconfig.core/META-INF/MANIFEST.MF
@@ -3,8 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Run Config Core Plug-in
 Bundle-SymbolicName: de.uka.ipd.sdq.workflow.launchconfig.core
 Bundle-Version: 5.2.0.qualifier
-Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core,
- de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder
+Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core
 Require-Bundle: de.uka.ipd.sdq.workflow;bundle-version="5.2.0",
  org.eclipse.debug.core;bundle-version="3.20.100"
 Bundle-ActivationPolicy: lazy
